requirements.txt
inflect>=0.2.5
librosa>=0.6.0
scipy>=1.0.0
Unidecode>=1.0.22
pillow
torchsummary
numba==0.48
matplotlib==2.1.0
tensorflow==1.15.2
numpy==1.13.3

